Refactor MiddlewareRepositorySession out of Crypto5MiddlewareRepositorySession.
