+++ src/main/java/org/mozilla/gecko/sync/middleware/Crypto5MiddlewareRepositorySession.java
@@ -43,17 +43,12 @@ import java.util.concurrent.ExecutorService;
 import org.mozilla.gecko.sync.CryptoRecord;
 import org.mozilla.gecko.sync.crypto.CryptoException;
 import org.mozilla.gecko.sync.crypto.KeyBundle;
-import org.mozilla.gecko.sync.repositories.InactiveSessionException;
 import org.mozilla.gecko.sync.repositories.NoStoreDelegateException;
 import org.mozilla.gecko.sync.repositories.RecordFactory;
 import org.mozilla.gecko.sync.repositories.RepositorySession;
 import org.mozilla.gecko.sync.repositories.RepositorySessionBundle;
-import org.mozilla.gecko.sync.repositories.delegates.RepositorySessionBeginDelegate;
 import org.mozilla.gecko.sync.repositories.delegates.RepositorySessionFetchRecordsDelegate;
-import org.mozilla.gecko.sync.repositories.delegates.RepositorySessionFinishDelegate;
-import org.mozilla.gecko.sync.repositories.delegates.RepositorySessionGuidsSinceDelegate;
 import org.mozilla.gecko.sync.repositories.delegates.RepositorySessionStoreDelegate;
-import org.mozilla.gecko.sync.repositories.delegates.RepositorySessionWipeDelegate;
 import org.mozilla.gecko.sync.repositories.domain.Record;
 
 /**
@@ -95,9 +90,8 @@ import org.mozilla.gecko.sync.repositories.domain.Record;
  * @author rnewman
  *
  */
-public class Crypto5MiddlewareRepositorySession extends RepositorySession {
+public class Crypto5MiddlewareRepositorySession extends MiddlewareRepositorySession {
   private KeyBundle keyBundle;
-  private RepositorySession inner;
   private RecordFactory recordFactory;
 
   public Crypto5MiddlewareRepositorySession(RepositorySession session, Crypto5MiddlewareRepository repository, RecordFactory recordFactory) {
@@ -182,13 +176,6 @@ public class Crypto5MiddlewareRepositorySession extends RepositorySession {
   }
 
   @Override
-  public void guidsSince(long timestamp,
-                         RepositorySessionGuidsSinceDelegate delegate) {
-    // TODO: need to do anything here?
-    inner.guidsSince(timestamp, delegate);
-  }
-
-  @Override
   public void fetchSince(long timestamp,
                          RepositorySessionFetchRecordsDelegate delegate) {
     inner.fetchSince(timestamp, makeUnwrappingDelegate(delegate));
@@ -230,93 +217,5 @@ public class Crypto5MiddlewareRepositorySession extends RepositorySession {
     }
     // Allow the inner session to do delegate handling.
     inner.store(rec);
-  }
-
-  @Override
-  public void wipe(RepositorySessionWipeDelegate delegate) {
-    inner.wipe(delegate);
-  }
-
-  @Override
-  public void storeDone() {
-    inner.storeDone();
-  }
-
-  @Override
-  public void storeDone(long storeEnd) {
-    inner.storeDone(storeEnd);
-  }
-
-  public class Crypto5MiddlewareRepositorySessionBeginDelegate implements RepositorySessionBeginDelegate {
-    private Crypto5MiddlewareRepositorySession outerSession;
-    private RepositorySessionBeginDelegate next;
-
-    public Crypto5MiddlewareRepositorySessionBeginDelegate(Crypto5MiddlewareRepositorySession outerSession, RepositorySessionBeginDelegate next) {
-      this.outerSession = outerSession;
-      this.next = next;
-    }
-
-    @Override
-    public void onBeginFailed(Exception ex) {
-      next.onBeginFailed(ex);
-    }
-
-    @Override
-    public void onBeginSucceeded(RepositorySession session) {
-      outerSession.setStatus(SessionStatus.ACTIVE);
-      next.onBeginSucceeded(outerSession);
-    }
-
-    @Override
-    public RepositorySessionBeginDelegate deferredBeginDelegate(ExecutorService executor) {
-      return this;
-    }
-  }
-
-  public void begin(RepositorySessionBeginDelegate delegate) {
-    inner.begin(new Crypto5MiddlewareRepositorySessionBeginDelegate(this, delegate));
-  }
-
-  public class Crypto5MiddlewareRepositorySessionFinishDelegate implements RepositorySessionFinishDelegate {
-    private Crypto5MiddlewareRepositorySession outerSession;
-    private RepositorySessionFinishDelegate next;
-
-    public Crypto5MiddlewareRepositorySessionFinishDelegate(Crypto5MiddlewareRepositorySession outerSession, RepositorySessionFinishDelegate next) {
-      this.outerSession = outerSession;
-      this.next = next;
-    }
-
-    @Override
-    public void onFinishFailed(Exception ex) {
-      next.onFinishFailed(ex);
-    }
-
-    @Override
-    public void onFinishSucceeded(RepositorySession session, RepositorySessionBundle bundle) {
-      outerSession.setStatus(SessionStatus.DONE);
-      next.onFinishSucceeded(outerSession, bundle);
-    }
-
-    @Override
-    public RepositorySessionFinishDelegate deferredFinishDelegate(ExecutorService executor) {
-      return this;
-    }
-  }
-
-  @Override
-  public void finish(RepositorySessionFinishDelegate delegate) throws InactiveSessionException {
-    inner.finish(new Crypto5MiddlewareRepositorySessionFinishDelegate(this, delegate));
-  }
-
-  @Override
-  public void abort() {
-    setStatus(SessionStatus.ABORTED);
-    inner.abort();
-  }
-
-  @Override
-  public void abort(RepositorySessionFinishDelegate delegate) {
-    this.status = SessionStatus.DONE; // TODO: ABORTED?
-    inner.abort(new Crypto5MiddlewareRepositorySessionFinishDelegate(this, delegate));
   }
 }
